+++ common/primitives/core/src/omni/mod.rs
@@ -14,31 +14,13 @@
 // You should have received a copy of the GNU General Public License
 // along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
 
-use async_trait::async_trait;
-use tokio::sync::mpsc;
+// This file contains the primitives definition for various omni-* services
 
-use executor_primitives::Intent;
+mod account;
+pub use account::*;
 
-/// Used to perform intent on destination chain
-#[async_trait]
-pub trait IntentExecutor: Send {
-	async fn execute(&self, intent: Intent) -> Result<(), ()>;
-}
+mod chain;
+pub use chain::*;
 
-pub struct MockedIntentExecutor {
-	sender: mpsc::UnboundedSender<()>,
-}
-
-impl MockedIntentExecutor {
-	pub fn new() -> (Self, mpsc::UnboundedReceiver<()>) {
-		let (sender, receiver) = mpsc::unbounded_channel();
-		(Self { sender }, receiver)
-	}
-}
-
-#[async_trait]
-impl IntentExecutor for MockedIntentExecutor {
-	async fn execute(&self, _intent: Intent) -> Result<(), ()> {
-		self.sender.send(()).map_err(|_| ())
-	}
-}
+mod intent;
+pub use intent::*;
